Add option to show/hide entries (#7)
Summary:

As beer lists grow, it is not necessary or desirable to show all beers in the beers.json file, yet removing that data would also not be good. This PR adds a visible key -- if visible is set to true, show the entry. Otherwise, don't show it. Also includes some README updates.