Support for labels/images in tap entries (#11)
Adds support to show images in the tap entry components. If an image is specified and is available, replaces the ABV/IBU/SRM block on the right, and the ABV/IBU/SRM data are presented at the bottom of the tap entry. Also adds conditional rendering for ABV/IBU/SRM. If these data are not available, then they are not shown.